Relocate change family screen file
