+++ composeApp/src/commonMain/kotlin/com/github/familyvault/ui/components/ConnectionError.kt
@@ -17,7 +17,7 @@ import com.github.familyvault.repositories.IFamilyGroupCredentialsRepository
 import com.github.familyvault.services.IFamilyGroupSessionService
 import com.github.familyvault.ui.components.typography.Paragraph
 import com.github.familyvault.ui.screens.LaunchingScreen
-import com.github.familyvault.ui.screens.main.ChangeFamilyGroupScreen
+import com.github.familyvault.ui.screens.main.familyGroupSettings.ChangeFamilyGroupScreen
 import familyvault.composeapp.generated.resources.Res
 import familyvault.composeapp.generated.resources.connection_error_button_content
 import familyvault.composeapp.generated.resources.connection_error_content
